// src/app/shared/models/booking.ts
export class Booking {
  id: string;
  nome: string;
  ultimoNome: string;
  cpf: string;
  dataNascimento: Date;
  email: string;
  nomeAcompanhante: string;
  ultimoNomeAcompanhante: string;
  cpfAcompanhante: string;
  dataNascimentoAcompanhante: Date;
  emailAcompanhante: string;
  cep: string;
  logradouro: string;
  pais: string;
  uf: string;
  localidade: string;
  telefone: string;
}